requirement.txt
streamlit==1.3.0
numpy==1.21.5
opencv-python-headless==4.5.5.62
Pillow==8.4.0